# COObject multivalue optimizations