--- a/Core/COObject+RelationshipCache.h
+++ b/Core/COObject+RelationshipCache.h
@@ -15,4 +15,10 @@
                             ofPropertyWithDescription: (ETPropertyDescription *)aProperty;
 - (void) removeCachedOutgoingRelationships;
 
+- (void) removeCachedOutgoingRelationshipsForCollectionValue: (id)obj
+								   ofPropertyWithDescription: (ETPropertyDescription *)aProperty;
+
+- (void) addCachedOutgoingRelationshipsForCollectionValue: (id)obj
+								ofPropertyWithDescription: (ETPropertyDescription *)aProperty;
+
 @end