--- a/Core/COPrimitiveCollection.h
+++ b/Core/COPrimitiveCollection.h
@@ -38,16 +38,27 @@
 - (BOOL)containsReference: (id)aReference;
 @end
 
+/**
+ * References are either COPath or any other object.
+ *
+ * COPath are treated as "tombstones" and hidden from the NSArray
+ * access methods (-count, -objectAtIndex:, etc.)
+ */
 @interface COMutableArray : NSMutableArray <COPrimitiveCollection>
 {
 @public
 	BOOL _mutable;
 	NSPointerArray *_backing;
-	NSMutableIndexSet *_deadIndexes;
+	/**
+	 * Array of integers, the ith entry gives the backing index for
+	 * "external" index i.
+	 */
+	NSPointerArray *_externalIndexToBackingIndex;
 }
 
 @property (nonatomic, readonly) NSPointerArray *backing;
 
+- (NSUInteger)referencesCount;
 - (id)referenceAtIndex: (NSUInteger)index;
 - (void)addReference: (id)aReference;
 - (void)replaceReferenceAtIndex: (NSUInteger)index withReference: (id)aReference;