The first 4 commits are just adding tests/benchmarks, the next 4 have the optimizations:

* rewrote handling of tombstones in `COMutableArray` to be faster. `-count` and `-objectAtIndex:` are now trivial, O(1).
* make `COUnsafeRetainedMutableArray` automatically remove duplicates, like NSOrderedSet. This lets us avoid the slow `-[COObject removeDuplicatesInValue:propertyDescription:]` every time a collection is modified.
* make a fast path version of `-commonDidChangeValueForProperty:` that just handles the updated elements of the collection. It's called: `-commonDidChangeValueForProperty:atIndexes:withObjects:mutationKind:`.

The benchmark I measured this with is TestLargeOrderedRelationsip (added in the first commit in this branch).

The benchmark incrementally adds 1000 OutlineItems to an OutlineItem with -addObject, then times various calls on `parent.contents` like `-count` and for/in enumeration.
Rewriting the tombstone code in COMutableArray made  `-count` ~30x faster when called on a collection with 1000 elements, and for/in enumeration ~5x faster.

The slowest part originally was building the collection of 1000 outline items by calling -addObject: in a loop, which was taking 10 seconds! Adding the fast path for `-commonDidChangeValueForProperty:` brought this down to 0.25 seconds (still kind of bad, but tolerable.)

(edit: all benchmarks re-run with 17aefef )
Before: 
```
2015-09-11 13:16:45.358 BenchmarkCoreObject[79618:1863789] 92.141032 us per iteration for 'Create OutlineItem and add to an NSMutableArray with -addObject:'
2015-09-11 13:16:56.758 BenchmarkCoreObject[79618:1863789] 11.399381 ms per iteration for 'Add OutlineItem to an OutlineItem with -addObject:'
2015-09-11 13:16:56.758 BenchmarkCoreObject[79618:1863789] 0.133991 us per iteration for 'Add OutlineItem to an NSMutableArray with -addObject:'
2015-09-11 13:16:56.759 BenchmarkCoreObject[79618:1863789] 0.459015 us per iteration for '-count on CoreObject array with 1000 elements'
2015-09-11 13:16:56.763 BenchmarkCoreObject[79618:1863789] 0.009000 us per iteration for '-count on NSArray with 1000 elements'
2015-09-11 13:17:00.693 BenchmarkCoreObject[79618:1863789] 3.930156 ms per iteration for 'for/in loop on CoreObject array with 1000 elements'
2015-09-11 13:17:00.698 BenchmarkCoreObject[79618:1863789] 4.337966 us per iteration for 'for/in loop on NSArray with 1000 elements'
```
After:
```
2015-09-11 13:15:34.698 BenchmarkCoreObject[79367:1862903] 155.786991 us per iteration for 'Create OutlineItem and add to an NSMutableArray with -addObject:'
2015-09-11 13:15:34.829 BenchmarkCoreObject[79367:1862903] 130.748034 us per iteration for 'Add OutlineItem to an OutlineItem with -addObject:'
2015-09-11 13:15:34.830 BenchmarkCoreObject[79367:1862903] 0.187993 us per iteration for 'Add OutlineItem to an NSMutableArray with -addObject:'
2015-09-11 13:15:34.830 BenchmarkCoreObject[79367:1862903] 0.014007 us per iteration for '-count on CoreObject array with 1000 elements'
2015-09-11 13:15:34.830 BenchmarkCoreObject[79367:1862903] 0.009000 us per iteration for '-count on NSArray with 1000 elements'
2015-09-11 13:15:34.970 BenchmarkCoreObject[79367:1862903] 139.808953 us per iteration for 'for/in loop on CoreObject array with 1000 elements'
2015-09-11 13:15:34.974 BenchmarkCoreObject[79367:1862903] 4.342973 us per iteration for 'for/in loop on NSArray with 1000 elements'
```

Edit: after removing relationship snapshotting:
```
2015-09-11 13:14:07.742 BenchmarkCoreObject[78716:1861487] 126.591027 us per iteration for 'Create OutlineItem and add to an NSMutableArray with -addObject:'
2015-09-11 13:14:07.791 BenchmarkCoreObject[78716:1861487] 48.901021 us per iteration for 'Add OutlineItem to an OutlineItem with -addObject:'
2015-09-11 13:14:07.792 BenchmarkCoreObject[78716:1861487] 0.321984 us per iteration for 'Add OutlineItem to an NSMutableArray with -addObject:'
2015-09-11 13:14:07.792 BenchmarkCoreObject[78716:1861487] 0.021994 us per iteration for '-count on CoreObject array with 1000 elements'
2015-09-11 13:14:07.793 BenchmarkCoreObject[78716:1861487] 0.009000 us per iteration for '-count on NSArray with 1000 elements'
2015-09-11 13:14:07.951 BenchmarkCoreObject[78716:1861487] 157.586992 us per iteration for 'for/in loop on CoreObject array with 1000 elements'
2015-09-11 13:14:07.955 BenchmarkCoreObject[78716:1861487] 4.539967 us per iteration for 'for/in loop on NSArray with 1000 elements'
```