@ericwa 2015-09-07
Argh, my indentation looks awful in github, here is what I'm using:
<img width="148" alt="screen shot 2015-09-07 at 1 51 02 am" src="https://cloud.githubusercontent.com/assets/239161/9711807/c8d4c224-5503-11e5-92b0-2a37653ee505.png">

Also, I know the last commit is ugly, I don't mind refactoring it more before merging if you want.

I think it is important to address that as soon as possible though (only processing collection changes, not re-checking all elements); I'm surprised we didn't do it before release because it only took a few hours.