@qmathe 2015-09-08
Thanks for this refactoring, it looks great :)

For the indentation, I'm using the same settings than you. However I have turned off Xcode automatic indentation and I tend to indent everything by hand. For Étoilé tabs + spaces combination to indent method keywords (and keep things aligned indepedently of the tab size), this doesn't work well with Xcode which always wants to turn spaces into tabs everyhere. So I tend to do everything by hand or use some more flexible editor.

I don't mind if the indentation is somewhat broken. This can be fixed later.